LUA - Advanced Missile Tracking
We've had this unique feature for a long time - AA Missiles that will seek a new target if their existing target dies while in-flight.

Well - it never worked right - but with the work on Projectiles, we've located the correct place to initiate the process - and know it works.   We'll need to watch it for any odd behavior we may have introduced - but we'll see - it has a certain amount of impact - since it can be busy when there are lots of SAM's in the air.