LUA - Effectutils - Building effects
As pointed out, the new perm. projectiles were malfunctioning for UEF Kennels, and Cybran engineers in general, when assisting structures other than factories.

Rather than re-enact the old bifurcated code which ran differently for repairs and assists, and had to check that, we simply warp the build projectile onto the assist target (which is correct) and the loop which moves the particle around, but exits when the unit is no longer being built - simply terminates.

This leaves the beam in precisely the right place, but without movement, until such time as the OnStopBuild() event happens, and the projectile is normally retrieved.