LUA: Improve Hotbuild keybind descriptions
These are at least an improvement from the human-unfriendly predecessors, and should imply that they're present to allow users to remap keys against build groups, rather than as indicators of what keys they are "supposed to use".

When I eventually show what key the user has mapped to each group over the construction menu icon of a unit in that group, it should be clearer than if these descriptions read something like `Hotbuild: W` for a unit with something other than W over its build icon.

I'll be listening for user feedback in the weeks after this goes live to see if I'm right about this.