LUA - Custom AI from map
A little known feature allows a specific AI to be forced on a start position - normally this field is blank, but we encountered a map that had a setting there - and of course, we no longer have the file it was looking for.

Correct this case to fall back to LOUD AI when it happens.