LOUD_MISC - Intel values on structure enhancments
As someone rightly pointed out - units which can enhance with radar/sonar/Omni are NOT affected by the Cybran Darkness.  Why ?  Because the blueprint says that, unenhanced, those units have 0 radar/sonar/OMNI - so the Darkness ignores them.

We have added a value of 1.1 to all of those units which can enhance themselves.   Not enough to be of any use, but enough for the Darkness to recognize them, and affect them if they are enhanced.