LUA - AI Engineer Attack moves
For some months now, LOUD has been using attack-moves for his engineers.   It's an interesting process, since reclaiming is hard to get right.   However..... on short jobs, like early in the game - we can't give up time to this behavior.

For now, I'll only use that kind of movement on paths that have multiple steps.