LUA - Guardpoint
Guardpoint forms the heart of a lot of AI platoon behaviors.   It's a platoon function that can direct a platoon to guard positions, or units (friendly & enemy),  with specific unit and structure requirements to trigger an end to the guard behavior, as well as time limits.

So - when you find it has a problem, you need to document it pretty carefully.  This update adds a great many outcommented LOG features that will detail the platoon behavior.   

Guardpoint was having just such a problem in it's movement routines - the platoon would immediately believe it was at the goal - rather than waiting for normal movement to get there.    This caused a situation where these platoons would 'beeline' directly to the final goal - I wasn't checking the range properly - greater than versus less than.    

Tightened up some of the timing loops to improve responsiveness.

Fixed.