LUA - AI Naval Defenses
The development of naval base static defense is something that's needed some focus for a while.   It's clear that early defenses are an ineffective method for defending a nascent naval facility - so we'll suspend any T1 defense at this time.

Alternatively, we'll boost priority of, and give some additional support to, the development of T2 naval defense, with the inclusion of amphibious T2 PD - and the ability for that builder to be assisted.

The economic triggers have also been adjusted.

The development of AA defense is now controlled by the AirStrengthRatio - and the development of T3 naval defenses is governed by the NavalStrengthRatio.