LUA - AI Outnumbered Expansion
The outnumbered condition has always stimulated LOUD to immediately begin expanding - now we want to carry that thru to his naval base as well - so that we get an almost immediate naval presence - and a very rapid 'fill-in' for those other close by start locations.

The minimum baserange is now reduced from 180 to 165.