LUA - T3 Power production
Moved the trigger points for T3 Pgen construction to encourage staying ahead of the needs.

- TrendOverTime goes from < 240 to < 260
- EnergyTrend goes from < 180 to < 200

We want to catch the potential start of a power crash a bit earlier - the idea being that we want another T3 gen - well under production - by the time the crash actually happens - if it does.    

Some will ask why even allow it to be possible ?   Isn't it easy just to overbuild power to make sure that doesn't ever happen ?

Yes - very easy in fact - but you see the cost clearly at the end of any game - win or loss - in the form of Energy Wasted.
   In fact those two metrics are perhaps THE most valuable in tuning AI economics -- eliminating waste, while avoiding crashes, is a tightrope to walk.