LUA - Outnumbered AI base expansion
Outnumbered AI will try to expand quickly - and to facilitate that we have split the Outnumbered functions between Land and Naval, in this fashion, we can use other metrics ( IMAPSize, MapWaterRatio, OutnumberedRatio) to either suppress or inflate the desire to expand particular expansions.

If the map is 20k, or less, and the WaterRatio(%) is less than 70% then we will modify the task priority - down if the WaterRatio is less than 50% - and up if it's  above.

This value is countered - partly - by the ratio by which the AI is outnumbered.  

This should result in a wide variety of responses to various map and team configurations.

The builders themselves, for both Land and Naval, have been modified with shorter search ranges.