LUA - GUARDPOINT exit conditions
When guarding - platoons will check for exit parameters based on structures and units - but we've added a 3rd option - and that is proximity to a base.

We noted this with a lot of the guard platoons - they'd continue to guard, once a base was started at their location, waiting for the normal structure and/or unit count triggers to go off.    While this is valid, I wanted a more immediate response - so they'll break from their guard cycle - and choose a new goal - or RTB. 

Now - another issue raised it's head - the platoon would pick the same location over again - not intended.   So - if the location picked is the same as where we're at already - we'll RTB. 

The code was supposed to be tracking the lastmarker but that needed some repair.

  The LOG will report this behavior so we can monitor just how many times it comes into play.

One minor repair to the feature where engineers pathing to start a new base will NOT use attack-move - it actually works properly now.