LUA - Teleportation updates
As we move to a better function for teleportation we need to fix some weaknesses in the teleport system

- extended range but at extended cost (max now 900 by default)

- we now pass both the unit bp and the teleport distance, into the actual teleport process (InitiateTeleportThread) to support this new calculation
- no more 'minimum time' to teleport, time will be determined by the unit build costs, the units buildpower (or the teleporting structure) - and modified by range - this will allow much longer range teleports, but at dramatically high costs. 

Note: depending upon the build power, some units will teleport very quickly - others not so well.

Units with dramatically improved build power will teleport faster simply because their energy flow rate (ala Evenflow) will be higher.

Added some new semaphores to more accurately indicate the current status of a teleport - to prevent double events, and to record the status of the initial burst charge - so that if a teleport fails, the burst charge remains unused - and doesn't have to be paid for again.  This fixes a long standing bug for the AI - where he would start a teleport,  and then abort it and walk.

New distance modifier which optimizes the efficiency of all teleportation at about 400 ogrids - but makes shorter ranges less efficient and longer ranges really costly.

A lot of additional logging as we work thru the data, so we can see the distance and cost values at work