LUA - Disappearing decals
Apparently, the process of putting resources on the map, at an early stage, 'locks' the decals in place.   Moving the creation to a point after creation of the 'brains' opens the possibility of having decals disappear (we don't know why).

So - in order to Create Resources() in it's usual sequence, we need to alert it to the use of MassPointRNG and MetalWorld, so that it can act accordingly - so we loop over the mods to see if they're present.

Although we loop over the mods again as we initialize the brains, it's for unit related things - and not gameplay choices.