LUA - ACU behavior
It was suggested that the mass extractor job for the ACU should persist for the entire game, and not just the first 5 minutes - let's give that a try.   

Also, the condition threat was different than the data threat value - so it could thrash - the ACU would pick the job up (because it had 30 threat or less) but then he would abandon it as his code would only allow upto 25 - bad.   Both are set to 35 now - why ? because I can.