LUA - Basic AI formations
Some of the merge sizes have proven to be a bit too large - so we've cut the maximum merge size in order to keep more platoons in action - likewise - the abort conditions trigger on those lower limits - which should drive the platoon to another choice.

Distress ranges and triggers have been changed for the smaller guard formations - so they won't cover quite as much range - but will respond to smaller threats.