LUA - AI CallForHelp parameters
Mex and Hydro platoons create a CallForHelp Thread that was only engaging when threat was 3 or higher - so - ignoring scouts and even LABS - this has been cut in half to 1.5 - which should mesh with some of the guard platoons now responding to threat levels that low.

Also moved the low mass trigger point for MEX upgrades up, just a bit - from .73 to .745.    This should make him just a bit less eager to upgrade when things are bad.

Slight increase in the number of instances for MEX and Mass Point guard formations.

Maximum formation size of T1 Guard platoons reduced to 4