LUA - Expanded LoopBuild functionality
Just as with assisting - you may wish to enable a loopbuilding engineer to have one set of requirements to start doing his job - but others that allow him to continue his building loop.

Currently, the code had NO break-off other than death - or not finding a suitable point.   

Two new builder data pieces allow the builder to have different values for both MASS and ENERGY when the builder loops.   This is useful in curtailing mass builders, especially in a MetalWorld scenario or normally - as overbuilding extractors can hammer the power systems.   

We now have the ability to break-off,  if the mass and energy values don't meet our needs - and turn a loopbuilder, into a single-shot builder, when things are economically too tight.

One other special concern here is entirely for mass building engineers - a break-off if the mass storage is approaching full.