LUA - AI Assisting
Taking advantage of the new assisting controls to permit my engineers who might be assisting building power, can continue to assist to a point where the econ stored is below the required amount necessary to start the job.

This permits more consistent, and urgent, assisting during periods of need for energy.    

You can see how the new values are added to an existing assist builder.