LUA - AI Mass builders
As we continue to examine the MW scenario - it was clear that there were no restraints on building extractors - it really was never conceived that we could have too many - well, you can.

This work puts a number of new conditions into play regarding building Mass extractors - primarily a condition that won't allow them until there is actually a deficit in storage, and some minimum requirements both in storage - and energy rate.  The energy rate is critical since we want to insure a strong - and steady E base before building more mass.

Also here, a small nudge in power requirements for mass extractor upgrades.

The goal is simple, curtail mass builds to keep more in tune with the ability of the E systems to provide power.

Broke up the builders into more choices, especially for the looping builders.   As they will now typically appear a little later than we've been used to - we want to make sure they're spreading out.

For the looping builders, you'll see the new loopbuilding values appear that alter how the builder will 'loop' his builds.