LUA - AI Energy builders
Still following the MW scenario - it was clear that we did not prepare LOUD to meet the energy demands of a rapidly growing eco base - especially with respect to having the E required.

Having more T1 options available, and more aggressive assisting helps - but this finishes that task.

There is considerably more emphasis on T1 power when needed, the number of engineers that can build T1 power goes up to 5, from 2.

The number of T1 pgens can now reach 80, up from 64, with more focus on early adjacency - this applies to T2 power as well - now having 9 positions, with a better sequence aim at more adjacency.

Reintroduced the T2 power build at Naval yards - but this is under watch for now.