LUA - AI engineers
A further adjustment to engineers and the use of 'attack move' - the ability is now part of the builder data.   So engineers can engage in that kind of movement, but by default they do not.

A tiny reduction in both the quantity of - and range of - mass reclaimers.

By default, we no longer build T1 transports in advance of actually needing them.

A logic flaw in the LandStrengthRatioGreaterThan function was reporting incorrectly.