LUA - SelfUpgrade Process
This is a fairly serious review.  We needed to move the reality of upgrading into the same realm as the reality of the economies, so that upgrades can flow naturally without crashing the eco.

This involves a better calculation of mass and energy trends required to upgrade something, and a review of how an upgrade impacts other upgrades waiting to happen.

For many units - the basic trend and efficiency triggers have been tightened up - those happen at the point those units are built.

The process itself, has been extensively revised and annotated, and the logging improved to allow pretty direct tracking of the decision process to upgrade, and it's impact.