LUA - ratios fix - divide by zero
The ratio formula could suffer from this if units were detected, but had no combat value - like air scouts for example.

Just added a default - note that .011 means you've seen this class of unit, but it has no value.   Compared to a result of 0.010 which means you have not seen this class of unit at all.