LUA - GUARDPOINT features
Addition of new 'merge' control feature that permits merging only with platoons of the same behavior AND the same builder.

A 2nd feature controls the start of guarding based on the largest distance trigger value (Unit or Structure), was previously always the Unit radius.

Better logging of process (out-commented)

Finer control over the AvoidsBases feature - was originally measuring distance from the platoon - not the marker or unit that the platoon was to guard.   This was unreliable over short distances an was triggered at 120.   This is now tied to the PointMin value in the builder data - allowing us to now guard points much closer to bases without generating a false exit.