LUA - AI Strength Ratio
The AI could mistakenly drop a strength ratio back to 0.011 (no enemy) when in fact it should have been 10 (total superiority), if the enemy had no units, or did not build any of that type.

We solve this by adding a time check, so that if 0 value is detected, after 10 minutes, we assume a 10 to 1 ratio, rather than not detected.