LUA - GUARDPOINT triggered exit
While moving to a point, the Unit and Structure triggers can fire - causing the immediate termination of the current point.    That works - but the thread that was moving the platoon was not being terminated, and that caused the new movement thread (created by picking another point) to be ignored.

The upshot was these platoons would get lost in limbo, unless merged with, or killed.

Good catch - again, frequently repeated observations, with easily reachable conditions, make finding this kind of logic issue at lot less work.   Testing GUARDPOINT, with all of it's various options, makes it more than trivial work.