LUA - FindPointMeetsConditions
This is a well-used function, so adding new features to it represents a lot of possibility.

In this case, after all the normal filtering is done, the final point list is normally sorted for distance, closest or furthest (and usually safest).   Now, with this option, we can have the list sorted for highest threat (we've already filtered out those with threat that is too high).  

I have added the 'MostThreat' option to the PointSort parameter of builders using GUARDPOINT, so the pointlist can be sorted on a threat basis.