LUA - AI Initial Air unit builders
We want the 'Initial' Transport builder to fire once, along the same lines as the 'Initial' Scout and Engineer builders.    We also want those builders to be removed once used.  

We set the ArmyNeedsTransport condition to TRUE, right at game start, to make that very clear.

Finally found the underlying cause of the overbuilt transport scenario - the added function ResetBrainNeedsTransport, was not being loaded for some of the builders that were responding to that condition.