LOBBY - Added the Auto-Rename Replay option
The option, to automatically rename replays according to map and timestamp is now found in the Interface section of the Options (from Main Menu).

I have also disabled the 'Medium' setting for Texture Detail.   While it can work for many, some of the later versions of DirectX are not entirely compatible with this, and it can lead to crashing.

Changed some of the text rollovers to better describe what the impact of them can be.