LUA - Engineer behavior
Engineers lead a dangerous life, and it's been aware of threat, at a build location, for a very long time - and for many years, I've intended to closely examine just exactly what levels of threat need to be used.

Let's just say - for a long time, 30 threat was the number - in reality, anything above 10 is assuredly lethal to a T1/T2 engineer.   So, we now use that value.

We'll continue to observe and refine that result.

Engineers will now RTB, or if they are looping engineers, another destination for their next build.

I also improved and fleshed out the highly detailed AI Debug for engineers.    This really let's you see the process of each engineer - but it's VERY detailed - use it lightly.