LUA - MovePlatoon function, RTB function
MovePlatoon keeps a copy of the original path - and removes completed waypoints from that, so that the original platoon can keep track of which part of the path it's on.

This is useful, since we can now properly maintain not only the final goal of the path - but the current stage.   This allows us to insert a threat check, during platoon travel, that can terminate pathed movement.

This was introduced to the GUARDPOINT behavior, for now, and it goes further, allowing GUARDPOINT to terminate the current mission, and either find another point - or begin an RTB process.

The RTB process has also been modifed, by introducing a MergeIntoNearbyPlatoon() so that RTB platoons can merge along the way with platoons using the same behavior.

Reduced the 'Stuck' period to 75 seconds from 90 and reduced the movement check cycle to 5 seconds, from 5.5