LUA - GUARDPOINT enhancements
As we continue to refine this behaviors features, we introduce the ability to Dialog it's behavior - I'll add this to the AI Debug process later - but for now it's stand-alone.

The key new feature is a threat check, performed by looking ahead to the current waypoint while moving.  In hindsight, while the original path was deemed 'safe' by the pathfinder - while underway - we could find threat blocks that we simply didn't acknowledge - and platoons would just stumble into them.

During path travel, GUARDPOINT will now look at the threat along the current path stage, and abort the move (and the current point) - forcing it into looking for another suitable point.