LUA - Engineer Structure Reclaim features
A bit of a chronic issue, we examined the 'reclaim' behavior of LOUD around his bases, and have come to the conclusion that the range is overly generous, using the operational range of the base the engineer is attached to (often 100+).

A more suitable range, and a way to specify the reclaim range directly, in the builder, was needed and we introduce it here.  You'll see ReclaimRange as a part of the builder data when we want to use this, and it will override the operations range of the manager that the engineer is attached to.