LUA - AirForceAILOUD - fighter behavior
Fighters are getting a bad reputation for throwing themselves at well defended positions in order to get a target - today, we'll move the threatcheckradius out to 128 from 96.

For smaller maps (20k and less) this will result in him using a larger IMAP block radius (1 versus 0), which may address the issue.  

Also correct some timing issues so they are skipped if we have to abort, resulting in a snappier response to the condition.