LOUD_UNITS - Cybran MegaRoach
Overall, it's balance was on the high side, so I have reseated it's role a bit more in the direction of MonkeyLord.

- HP reduced 121K >> 105.5K
- Regen reduced 100 >> 65
- Both Mass & Energy cost up
- Radar and vision slightly reduced
- Vet requirements +7%
- Turret speeds reduced

This has been a long time coming, and as a Cybran fanboi, you can blame me for not wanting to drag this unit into it's class - but it was overdue.