LUA - AI Mass Point Share
The AI uses the 'Mass Point Share' value to initiate a trigger in it's posture, from Skirmish mode, over to Assault mode.  Originally, this was based upon a simple value based upon # of players and # of mass points on the map.  This value was originally capped at 10 plus the # of players.

We'll now cap that value at 8 plus, to encourage an earlier posture change, especially on mass heavy maps where there may not be many players.