LUA - AI Outnumbered Cheat
We've been encouraged by the limited changes we have in place for the Outnumbered AI - it helps, a little, to overcome the greater options that a larger team has.

Today, we're introducing a subtle new cheat, just for the Outnumbered AI - in the form of a build rate boost.   No additional resources, but the AI will have an enhancement to his build rate, on top of his existing cheat.    This will be limited to the lesser of either the Outnumbered Ratio - or his standard cheat.

For example, an Outnumbered AI facing 3 against 2 - @ 1.2, will get his build rate increased by another 1.2x.

This is a work in progress, so we'll monitor it for now and see how it fares.