LUA - Weapon State logging
I've added the entire gamut of log entries to Defaultweapons.   This permits tracking of every weapon state, and was instrumental in highlighting the impact of 'busy' state weapons, especially on those units with lots of weapons.

At the moment, they are commented out - but they should likely become an option in the AI Debug utility.