EFFECTS - Ver 3.2
This is a minor update, the addition of the Miasma cloud effect, which I have created for the Miasma Field (AEON BO ACU).

This effect is simply the Miasma artillery effect, but set to play continuously - and be much closer to the ground (mostly concealed) - the idea being that it's  a corrosive 'acid' effect at surface level.