LUA - BUFFFIELD changes
Some structural work.

Removing circular reference to Owner - replacing it with the Sync.id - and doing the lookup when needed.  This permits us to log details of the bufffield if needed, without overflow from the owner entity.

Carrying in the field radius from the blueprint.

Carrying in the VisualScale parameter from the blueprint - and using it when we create the effects.

Passing thru the bp in follow on function calls, saving repeated lookups