BO ACU - AEON ACU - 1st look
This is the first completed ACU from the BO review.

There is lots to see here so I expect the rumor mill can stop feeding on itself - but most notably - a substantial change to the mobility of the ACU.

Compared to the standard ACU - the BO ACU starts will 245% more HP - and as such - his nominal movement speed begins at 1.45 - down from the 1.7 of the standard ACU.    Likewise, his acceleration is set to require 3 seconds to reach top speed.

The engineering lines directly affect mobility.

Combat engineering will impair mobility as more armor is belted on - with a scaled penalty of 4, 8 and then 12% respectively.  At the top end, max speed comes in just under 1.3.

Standard engineering, conversely, increases mobility with a scaled bonus of 6,12 and 18% - with regen bonuses on the 2nd and 3rd upgrade.  Fully upgraded, along this line, the ACU regains his movement speed of 1.7 as a result.

All engineering upgrade costs are normalized across the stages and this represents a reduction in cost for the standard engineering upgrades.    Build power bonuses for each have been adjusted accordingly, the maximum build power now tops out at 160 (same as fully upgraded factories) - and there are minor variances in the resource bonuses, as we have gone from fixed amounts, to a multiplier based upon the base amount.

Basic regen has increased by 50%.

The biggest amount of work centers on the weapon systems.   Every weapon and auxiliary system has been reviewed and normalized, where possible, with those already existing in the game.   In general, almost all the weapon upgrades have decreased dramatically in cost - especially at the top end.

Have at it.