LUA - Teleportation
Two revisions today - a doubling of the maximum range for teleportation.   This will, obviously, vastly increase the charge required to do so - but you can.

Second - the initial energy cost, which is the minimum E storage required to start the teleport, is now removed from the charge cost - reducing the E load for the teleport itself.