LUA - Orders panel expands to 16 from 12
The number of available orders has been just sufficient up until now, with 12 spots.   For some units, this left only 2 available spots for 'ability' buttons and special toggles.

This work will revise, and resize that panel, to permit up to 16 slots, with some fixed additions planned for the Attack-Move order, and the fixed positioning of the OverCharge ability (was variable), will now appear in the 4th slot of the 1st line (next to the other Attack related orders).

The background for this work came from Domino Mod Support, which introduces a flexible sized panel that is ultimately, capable of up to 24 orders.

Ability definitions have been altered to permit both Slot selection, and script, to be passed directly from the unit blueprint - while still supporting the abilitydefinition table if present.

While not yet complete, some support work for the Attack-Move order is included here.

Some preliminary support for Hotkey labels is now in place as well, but this feature too - is not yet implemented.