LUA - Unit Detail info, Buff tweaks, teleportation
Updated unit regeneration to actually be gleaned from the buff itself, rather than relying upon iterating over the blueprints to find values in the enhancement.

So, when a buff happens, the regen value is updated on the unit itself - and passed over to the UI in the selectedinfo.lua module.

Some updates to buffs.

Teleport cost is reduced by the pre-charge amount