LUA - UI - Unit and Shield regen & Health buffs
Unit regeneration, and Shield Strength and regeneration were values that were NOT immediately available to the UI - they had to be calculated by looking at the kills, and iterating over the blueprints to pick up any buffs that may have occured.   This is problematic when the blueprint has incorrect values - and it's slow for the UI.

When a buff of that sort is applied, we now record the updated value in the STATS of the unit, so that it is directly retrievable without that work.