LUA, LOUD_MISC - Bad scores
The -1.@IND problem has been narrowed down to some strange 'event' which we don't seem to be causing ourselves.

This code work will trap it - and report it - when it happens - but cease reporting it - when the event ends and the economy incomes return to normal.   The economy totals (reclaimed and produced) however, never recover.