LUA - Reclaim values
An interesting investigation - due to poorly placed decimal point,  the time required to reclaim mass and energy, from props, has, in many cases been accelerated by as much as 10x.

The reclaim 'time' is built from the amount of mass and/or energy to be reclaimed - and applying the buildpower of the reclaimer to it.   Well, while the reclaim 'time' of a life unit was totally normal (a unit), for props (trees, rocks and wrecks), it was factored incorrectly.

In testing - putting the reclaim rates at equal to the build rate, is going to be a rather obvious change - so we'll try a compromise here, allowing that kind of reclaim to operate at 2x the normal build rate.  Of course, we also fixed the original decimal point issue, and adjusted for energy reclaim being at 10x compared to mass reclaim (same as Evenflow).