LUA - UI tests - unit list breaks on low resolution
The scrolling unit list, brought up when building something, breaks - when you select FULL SCROLL - leaving behind a ghost image , and not presenting the final item choice (or hiding it behind the scroll buttons).

The selection process works just fine if you step to the end, with single clicks.

The issue can be reset, by doing a FULL LEFT SCROLL (back to the start) of the list.

This issue seems to pertain primarily to low resolutions - but is related to the number of 'break' icons that are part of the list - not counted or the pixel count is not right.

This is just an investigation, I didn't really do anything but factor the code for my readability - and to install some log entries for later study.