BrewLAN - Mines
This is a general rebalance of mines based upon (but not quite the same) the same energy formulae as artillery utilizes.    This gives us a solid foundation for altering mine damage and AOE radius.

I have tried to position the 3 mine types so that they are reasonably affordable - and effective - within the tier they appear.   It's obviously open for re-interpretation at some point, but this should lower the bar for wider usage.