BrewLAN - Field Engineers
The response to the FE has generally been underwhelming - and I have to assume that's primarily due to it's value proposition.    I've nudged that slightly, and in conjunction with the mine changes, I hope to see more usage.

Most of the changes revolve around other engineer related improvements (showing range ring and intel overlays) - but there are a few stat changes in some of  them.   Those with shields have been updated with the shield E usage formula.