LUA - Expansion of BUFF abilities
The existing 'MoveMult' buff is too coarse for some of things I'd like to do, so I have introduced discrete buffs for Speed, Acceleration and Turning. 

I have new buffdefinitions that will utilize that feature to provide an 'agility' buff - which enhances acceleration and turn speed, but not top speed.