BO ACU - First Release of new ACU configurations
This is the first full push of the completed BO ACU options, fully balanced for LOUD.

There are a LOT of changes here, so I won't go into much detail, but every weapon and auxiliary option has been balanced against the LOUD universe, so that in most cases, cost is substantially reduced, and performance of the weapons dovetails with existing weapons found elsewhere.

I have updated the descriptions accordingly - but it's important to realize that so much is involved here that I may have overlooked many things.  Fingers crossed.