LUA - BuffFields
Expansion of the Bufffield blueprint to permit passing along visual data as part of the blueprint (previously hidden in class data)

The attachbone can now be specified on the unit blueprint where the bufffield is invoked.   This permits them to be re-used with different parameters.

Removed a Bufffield that was being defined in SeraphimWeapons module