BrewLAN - Sera Optical Tracking
The Optical Tracking device uses code very similar to the Eye of Rhianne - and as such, we consider it to operate on a very similar basis - teleported entities.

I wanted to bring anti-teleport into that unit, and prepare the core code to accept the feature set from the BrewLAN unit, which allows attachment to enemy units.