LUA - UNITS - fix incorrect value, logging
Enhancement was passing the wrong mass value to the UI

CDREnhance was not signalling it's status thru data - the Upgrading flag is now turned on, and the CurrentBuildOrder will read 'Enhance'.   Turn on DesiresAssist for upto 3 assistees, and report itself as the UnitBeingBuilt.

Resequenced when sound events trigger - after the consumptions are updated, and the enhancement effects begin.