LUA - ParseIntelThread
This being one of the most intensive processes in LOUD, it's important that we periodically review it.  It has been, recently, just a trifle 'too' hot.   This has, in some circumstances, caused a crash - so, were going to back it off a little - and do some optimization.

We also wanted to clean up some of the variable names for clarity.