LUA - localizing distance functions in sorts
Sorting data is sometime very intensive - and sometimes not - and sorting a table, by distance, is a fairly common process.   Now - the distance functions are usually upvalues to that process, which makes them a trifle slow - and bringing the distance function, into the sort can speed that up quite a bit - but it's really dependent upon just how big the table is - the bigger, the more benefit.   

This is, of course, offset by the extra memory required to elevate the distance function to a local, within the sort - and the extra garbage that needs to be relieved after the sort is complete, but at very least - it's a neutral impact - and the gain on the large sorts is well worth it.