LUA - Data reductions
Some wider applications of data reduction, in reducing iterating over tables, by localizing the specific item.

We're also doing some cleanup in this work, organizing the locals we use - in one place - and arranging them for easy viewing.   This often aids us in re-using variables and reducing the number of locals we need to create in a function.

This also gives the opportunity to relabel variables to more accurate purposes or more consistency between functions in the same module.