LUA - AI Utilities
A lot of variable cleanup underway now - with a more focused eye upon the data dictionary, and the variable table.   This directly impacts memory usage, as we'll be looking for not only redundant variables, but ways in which we can re-use them, to reduce the creation of additional locals in temporary usage.

This impacts a lot of utility functions, targeting aids and behavioral functions.