LUA - AI Debug features
Recently, I've been adding more data to my AI Debug work, so that it's easy to track individual things within the AI - and see the decision tree, in specific operations.    To do that, I've recently been adding the 'BuilderInstance' to the platoon names that appear, either visually, or in the log.