LUA - AI Debug features
Expanding upon the previous commit, we've expanded the detail in the Guardpoint logging to better reveal the behavior of the various 'flavors' of GuardpointAI - and add the Instance value to the platoon name so that it's now possible to trace an individual platoon thru it's decision process.

As we begin to re-write Guardpoint (in all it's flavors) this is an excellent opportunity to get into the 'nuts and bolts' of their process - and see the inner dialog.