LUA - Movement triggers
Optimizing the code for waypoint movement triggers and making the detection cycle dynamic with range to the trigger, to reduce the amount of tests.

Expanded the inclusion of the code to direct specific movement when the cycle value gets to high - this has been effective in reducing the incidence of stuck and dead platoons.