LUA - Bufffield effects
Bufffields and the effects that indicate their effect, were not functioning correctly, often leaving effects on, and sometimes the buff as well, after having left the field.

I also wanted to confirm the impact of stacking/replacing and the use of a duration versus a simple buff - and then a de-buff when having left the field.   This turned out to be the much lighter method, since we just use a semaphore to indicate if the buff is already present or not - rather than trying to reapply the buff in a given period.