LUA - Implemented WaypointSlackDistance
The WaypointSlackDistance (the distance at which a platoon recognizes that it's just about at a waypoint) can now be implemented at the platoon data level - rather than relying upon a default value. 

This is very useful for very large formations that may require a larger value simply due to the size of the formation.

Some minor reductions in MergeLimit, and DistressRange as well.