LUA - Builders - PriorityFunction
The PriorityFunction of builders is a very useful tool for adjusting priorities of a building, on the fly, with custom code.   The PlatoonFormManager, however, only reviewed these functions when a base changes it's status (Primary or not).

By carrying the function thru, from the Builders master table, to the builders used in each bases PFM, we can invoke them with each cycle of the PFM

Improved a great deal of Platoon Dialog logging