LUA - Formation (Crash bug)
Finally located the random crash bug - in Formations.  Was being caused when a mixed formation (subs and surface ships) had an unlikely distribution of units (ie. - more subs than surface vessels).

In those cases, normally, the SUBS are placed into the formation first - and then removed from the total units before placing the surface vessels, in the same formation, over top of them.   The crash would occur, because the subs were not removed from the total, and the code was trying to endless place units, into the formation, which did not exist.